Check for existing polygons when drawing features
When attempting to draw premises and pickup area polygons, the SDK will now check to see if a feature with that `id` exists before drawing. If it exists, it will be removed and redrawn.

This allows for the flexibility of changing polygons if necessary (though that would be rare) but it also allows for the `update` call to accept a `data.site` property with `premises_coordinates` and `pickup_areas` features without drawing multiple layers for the same feature.