Map Customization Options
We now support customizing maps via a configuration object.

*Note: `showControls` has been removed in favor of this more flexible configuration approach.*

If you are using Google Maps, you can pass any of the options that the Google Maps SDK supports and it will be passed directly to their map constructor.

If you are using Mapbox GL Maps, we currenly only support a `showNavigationControls` directive. (We can add more configuration keys as needed)